Update Main.py
In this verison new section was cleaned - DefinePopParamForMiddleSteps function. Now it gives the same results as the old version, except last step - population merge generate some demographic changes that do not match. Bm and x_last are still there, but should not be necessary, only stayed there so the last function works.